+++ crates/python/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.19.4-rc.0
+
+### Fixes
+
+- ExecutionOptions::default() now respects defaults used in ExecutionOptionsBuilder::default() (#499)
+
 ## 0.19.3
 
 ### Fixes